requirements.txt
Django==2.0.6
django-configurations==2.0
django-extensions==2.0.7
raven==6.9.0
pytest-django==3.2.1
factory-boy==2.10.0
